+++ tests/pipe/3pipes_reserve_producer_consumer.cpp
@@ -55,7 +55,7 @@ int test_main(int /*argc*/, char*[] /*argv*/) {
         { WI, WI },
         [=] (auto group) {
           // Use a sequential loop in the work-group to stream chunks in order
-          for (int start = 0; start != N; start += WI) {
+          for (int start = 0; start != (int)N; start += (int)WI) {
             /* To keep the reservation status outside the scope of the
                reservation itself */
             bool ok;
